Make the following changes:
,

Find this element:

  <Sidebar variant="sidebar" collapsible="icon" className="border-l" side="right">

with the text content:

    InitiativeToggle Sidebar Initiative TrackerTurn OrderRound: 1No participants in initiative.Next Turn Add ParticipantNameInitiativeTypePlayerEnemy Add to Initiative

Relevant files:

    - https://6000-firebase-studio-1747883817491.cluster-joak5ukfbnbyqspg4tewa33d24.cloudworkstations.dev/_next/static/chunks/src_228eb5c5._.js
    - https://6000-firebase-studio-1747883817491.cluster-joak5ukfbnbyqspg4tewa33d24.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the initiative tracker, change the name to Turn Tracker. At the bottom of the sidebar make a footer that has the button for Start Combat. When combat starts, replace the start combat button with the Next Turn and End Combat buttons in one row.