Make the following changes:
,

Find this element:

  <InitiativeTrackerPanel currentParticipantIndex=0 roundCounter=1>

with the text content:

    Turn OrderRound: 1 [data-radix-scroll-area-viewport]{scrollbar-width:none;-ms-overflow-style:none;-webkit-overflow-scrolling:touch;}[data-radix-scroll-area-viewport]::-webkit-scrollbar{display:none}  10Jurgen(HP: 10)(AC: 10)player 10Evil Guy(HP: 10)(AC: 10)enemy 10Friendo(HP: 10)(AC: 10)ally

Relevant files:

    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_a31e2e15._.js
    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the combatant cards, make the name more bold. When the combatant is active or it is their turn, instead of an orange background just give them a 2 px orange border. Change the player and enemy badges to rounded rectangles. Change player to green and ally to Blue. Make the badges first letter capitalized.