Make the following changes:
,

Find this element:

  <InitiativeTrackerPanel currentParticipantIndex=0 roundCounter=1>

with the text content:

    Turn OrderRound: 1 [data-radix-scroll-area-viewport]{scrollbar-width:none;-ms-overflow-style:none;-webkit-overflow-scrolling:touch;}[data-radix-scroll-area-viewport]::-webkit-scrollbar{display:none}  10Gorok(HP: 10)(AC: 10)Player

Relevant files:

    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_a31e2e15._.js
    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

For the combatant card, put the trash can in the top right in line with the name and add a three dots icon to the bottom right that will open a popover with the option to Rename or to Change Token (upload new). Make the icon light gray with no background and when hovered make it white.