Make the following changes:
,

Find this element:

  <CircleUserRound x=3 y=3 width=24 height=24 color="hsl(var(--primary-foreground))" strokeWidth=1.5 className="pointer-events-auto">

Relevant files:

    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_a31e2e15._.js
    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_a31e2e15._.js
    - https://6000-firebase-studio-1748124699058.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the token label text outline to be black. Make it thicker. The token label text drop shadow should be a blurry shadow like bg-shadow-md or whatever, the type of shadow on cards, not a hard drop shadow that just looks like more text. I also want this kind of card shadow on the token itself to help visually separate it and the text from the background.