Make the following changes:
,

Find this element:

  <FloatingToolbar activeTool="select" selectedColor="#FF0000" backgroundImageUrl="/default-maps/Glade.png" showGridLines=true showAllLabels=true measurement={"type":null} backgroundZoomLevel=1 canUndo=false canRedo=false escapePressCount=0 toolbarPosition="top">

Relevant files:

    - https://6000-firebase-studio-1748401402223.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_d11e1775._.js
    - https://6000-firebase-studio-1748401402223.cluster-hf4yr35cmnbd4vhbxvfvc6cp5q.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

Change the move toolbar icon colors to blue background and border with white icon and on hover just darken the background. For the reset board button make the background red and icon white and just darken the background a little on hover. Also add a divider between the reset board button and undo/redo buttons.